salloc_common
